question utils
