+++ frontend/src/utils/QuestionUtils.ts
@@ -1,0 +1,37 @@
+import React from "react";
+import { DropTargetMonitor } from "react-dnd";
+import type { XYCoord } from "dnd-core";
+
+export interface DragItem {
+  index: number;
+  id: string;
+  type: string;
+}
+
+export const compute = (
+  dragIndex: number,
+  hoverIndex: number,
+  previewRef: React.RefObject<HTMLDivElement>,
+  monitor: DropTargetMonitor<DragItem, void>,
+): boolean => {
+  if (dragIndex === hoverIndex) {
+    return false;
+  }
+  const hoverBoundingRect = previewRef.current?.getBoundingClientRect();
+  if (hoverBoundingRect) {
+    const hoverMiddleY = (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
+    const clientOffset = monitor.getClientOffset();
+    const hoverClientY = (clientOffset as XYCoord).y - hoverBoundingRect.top;
+
+    if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
+      return false;
+    }
+    if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
+      return false;
+    }
+
+    return true;
+  }
+
+  return false;
+};
